Audio drivers rework: - nuked QMidiEventQueue, incoming events are now supposed to be timestamped on the fly and added to the internal synth's queue - common timing-related methods and fields moved to abstract AudioStream class - audio drivers adopted to the new design - removed from WinMM driver those paranoid x-run checks (moreover, incorrect for Vista+) - more code cleanup

+++ mt32emu_qt/src/MasterClock.cpp
@@ -75,7 +75,7 @@ MasterClock::~MasterClock() {}
 // Why it's protected in QThread is a bit of a mystery...
 class QtSucks : public QThread {
 public:
-	static void usleep(int usecs) {
+	static void usleep(ulong usecs) {
 		QThread::usleep(usecs);
 	}
 };
